+++ src/tui/log_pane.rs
@@ -7,6 +7,11 @@ pub(crate) struct LogPane {
     log_db: LogDb,
     logger: Logger,
     rx_logs: Receiver<LogMessage>,
+    vertical_scroll_state: ScrollbarState,
+    horizontal_scroll_state: ScrollbarState,
+    vertical_scroll: usize,
+    horizontal_scroll: usize,
+    current_frame_area: Rect,
 }
 
 impl Default for LogPane {
@@ -18,27 +23,36 @@ impl Default for LogPane {
             log_db: Default::default(),
             logger,
             rx_logs,
+            vertical_scroll_state: ScrollbarState::default(),
+            horizontal_scroll_state: ScrollbarState::default(),
+            vertical_scroll: 0,
+            horizontal_scroll: 0,
+            current_frame_area: Rect::ZERO,
         }
     }
 }
 
 impl LogPane {
     pub fn render(&mut self, frame: &mut Frame, area: Rect, in_focus: bool) {
-        let logs = self.log_db.latest_logs(self.log_level());
-        let mut list_items: Vec<ListItem> = vec![];
+        let logs = self.log_db.all_logs(self.log_level());
+        let mut lines: Vec<Line<'_>> = vec![];
         for msg in logs {
             match msg {
                 LogMessage::Error(s) => {
-                    list_items.push(ListItem::new(s.as_ref()).red());
+                    lines.push(Line::from(s.as_ref()).red());
                 }
-                LogMessage::Warn(s) => list_items.push(ListItem::new(s.as_ref()).yellow()),
-                LogMessage::Info(s) => list_items.push(ListItem::new(s.as_ref())),
-                LogMessage::Debug(s) => list_items.push(ListItem::new(s.as_ref()).cyan()),
-                LogMessage::Trace(s) => list_items.push(ListItem::new(s.as_ref()).blue()),
+                LogMessage::Warn(s) => lines.push(Line::from(s.as_ref()).yellow()),
+                LogMessage::Info(s) => lines.push(Line::from(s.as_ref())),
+                LogMessage::Debug(s) => lines.push(Line::from(s.as_ref()).cyan()),
+                LogMessage::Trace(s) => lines.push(Line::from(s.as_ref()).blue()),
             }
         }
+        let content_len = lines.len();
 
-        let list = List::new(list_items);
+        self.vertical_scroll_state = self.vertical_scroll_state.content_length(content_len);
+        self.horizontal_scroll_state = self
+            .horizontal_scroll_state
+            .content_length(self.log_db.longest_message());
 
         let block_border_symbol = if in_focus {
             symbols::border::PLAIN
@@ -47,12 +61,34 @@ impl LogPane {
         };
 
         let log_block = Block::bordered()
-            .title(format!("Log Level: {}", self.log_level()))
+            .title(
+                format!(
+                    "Log Level: {}, showing {content_len} msgs",
+                    self.log_level()
+                )
+                .bold(),
+            )
             .border_set(block_border_symbol);
-
-        let log_widget = list.block(log_block);
-
-        frame.render_widget(log_widget, area);
+        let paragraph = Paragraph::new(lines)
+            .gray()
+            .block(log_block.gray())
+            .scroll((self.vertical_scroll as u16, self.horizontal_scroll as u16));
+        frame.render_widget(paragraph, area);
+        frame.render_stateful_widget(
+            Scrollbar::new(ScrollbarOrientation::VerticalRight)
+                .begin_symbol(Some("↑"))
+                .end_symbol(Some("↓")),
+            area,
+            &mut self.vertical_scroll_state,
+        );
+        frame.render_stateful_widget(
+            Scrollbar::new(ScrollbarOrientation::HorizontalBottom)
+                .thumb_symbol("🬋")
+                .begin_symbol(None)
+                .end_symbol(None),
+            area,
+            &mut self.horizontal_scroll_state,
+        );
     }
 
     pub fn get_logger_clone(&self) -> Logger {
@@ -75,5 +111,69 @@ impl LogPane {
 
     pub(crate) fn log_level(&self) -> LogLevel {
         self.logger.verbosity()
+    }
+
+    pub(crate) fn scroll_to_start(&mut self) {
+        self.vertical_scroll = 0;
+        self.vertical_scroll_state = self.vertical_scroll_state.position(self.vertical_scroll);
+        self.horizontal_scroll = 0;
+        self.horizontal_scroll_state = self
+            .horizontal_scroll_state
+            .position(self.horizontal_scroll);
+        self.log_db.unfreeze();
+    }
+
+    pub(crate) fn scroll_to_end(&mut self) {
+        self.vertical_scroll = self.log_db.len() - 1;
+        self.vertical_scroll_state = self.vertical_scroll_state.position(self.vertical_scroll);
+        self.log_db.freeze();
+    }
+
+    pub(crate) fn scroll_down(&mut self) {
+        self.vertical_scroll = self.vertical_scroll.saturating_add(1);
+        self.vertical_scroll_state = self.vertical_scroll_state.position(self.vertical_scroll);
+        self.log_db.freeze();
+    }
+
+    pub(crate) fn scroll_up(&mut self) {
+        self.vertical_scroll = self.vertical_scroll.saturating_sub(1);
+        self.vertical_scroll_state = self.vertical_scroll_state.position(self.vertical_scroll);
+        if self.vertical_scroll == 0 {
+            self.log_db.unfreeze();
+        }
+    }
+
+    pub(crate) fn scroll_left(&mut self) {
+        self.horizontal_scroll = self.horizontal_scroll.saturating_sub(1);
+        self.horizontal_scroll_state = self
+            .horizontal_scroll_state
+            .position(self.horizontal_scroll);
+    }
+
+    pub(crate) fn scroll_right(&mut self) {
+        self.horizontal_scroll = self.horizontal_scroll.saturating_add(1);
+        self.horizontal_scroll_state = self
+            .horizontal_scroll_state
+            .position(self.horizontal_scroll);
+    }
+
+    pub(crate) fn scroll_page_up(&mut self) {
+        let window_height = self.current_frame_area.height;
+        let lines_in_window = window_height.saturating_sub(2);
+        for _ in 0..lines_in_window {
+            self.scroll_up();
+        }
+    }
+
+    pub(crate) fn scroll_page_down(&mut self) {
+        let window_height = self.current_frame_area.height;
+        let lines_in_window = window_height.saturating_sub(2);
+        for _ in 0..lines_in_window {
+            self.scroll_down();
+        }
+    }
+
+    pub(crate) fn set_current_frame_area(&mut self, area: Rect) {
+        self.current_frame_area = area;
     }
 }