+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [unreleased]
 
+### Added
+
+- Add scrollbar and navigation for log pane
+
 ## [0.2.0]
 
 ### Changed